Restructure redux folder location
